+++ premake5.lua
@@ -526,24 +526,13 @@ function create_tests()
     
     group "benchmarks"
         
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_camera", include_tests="omni.isaac.benchmarks.tests.test_benchmark_camera.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_physx_lidar", include_tests="omni.isaac.benchmarks.tests.test_benchmark_physx_lidar.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_real_time_factor", include_tests="omni.isaac.benchmarks.tests.test_benchmark_real_time_factor.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_robots_nova_carter_ros2", include_tests="omni.isaac.benchmarks.tests.test_benchmark_robots_nova_carter_ros2.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_robots_nova_carter", include_tests="omni.isaac.benchmarks.tests.test_benchmark_robots_nova_carter.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_robots_o3dyn", include_tests="omni.isaac.benchmarks.tests.test_benchmark_robots_o3dyn.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_ros_camera", include_tests="omni.isaac.benchmarks.tests.test_benchmark_ros_camera.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_rtx_lidar", include_tests="omni.isaac.benchmarks.tests.test_benchmark_rtx_lidar.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_rtx_radar", include_tests="omni.isaac.benchmarks.tests.test_benchmark_rtx_radar.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_scene_generation", include_tests="omni.isaac.benchmarks.tests.test_benchmark_scene_generation.*"})
-        define_ext_test_experience("omni.isaac.benchmarks", {suite = "omni.isaac.benchmarks", name_override="benchmark_sdg_generation", include_tests="omni.isaac.benchmarks.tests.test_benchmark_sdg_generation.*"})
-        
-        python_sample_test("tests-standalone_benchmarks-benchmark_camera", "standalone_examples/benchmarks/benchmark_camera.py", "--test")
-        python_sample_test("tests-standalone_benchmarks-benchmark_robots_nova_carter_ros2", "standalone_examples/benchmarks/benchmark_robots_nova_carter_ros2.py", "--test")
-        python_sample_test("tests-standalone_benchmarks-benchmark_robots_nova_carter", "standalone_examples/benchmarks/benchmark_robots_nova_carter.py", "--test")
-        python_sample_test("tests-standalone_benchmarks-benchmark_rtx_lidar_rotary", "standalone_examples/benchmarks/benchmark_rtx_lidar.py", "--test -n 8 --lidar-type Rotary")
-        python_sample_test("tests-standalone_benchmarks-benchmark_rtx_lidar_solid_state", "standalone_examples/benchmarks/benchmark_rtx_lidar.py", "--test -n 8 --lidar-type Solid_State")
-        python_sample_test("tests-standalone_benchmarks-benchmark_sdg", "standalone_examples/benchmarks/benchmark_sdg.py")
+        python_sample_test("tests-standalone_benchmarks-benchmark_camera", "standalone_examples/benchmarks/benchmark_camera.py", "--num-frames 1 -num-cameras 2")
+        python_sample_test("tests-standalone_benchmarks-benchmark_robots_nova_carter_ros2", "standalone_examples/benchmarks/benchmark_robots_nova_carter_ros2.py", "--num-frames 1 -num-robots 2 --enable-3d-lidar 1 --enable-2d-lidar 2 --enable-hawks 1 ")
+        python_sample_test("tests-standalone_benchmarks-benchmark_robots_nova_carter", "standalone_examples/benchmarks/benchmark_robots_nova_carter.py", "--num-frames 1 -num-robots 2")
+        python_sample_test("tests-standalone_benchmarks-benchmark_rtx_lidar_rotary", "standalone_examples/benchmarks/benchmark_rtx_lidar.py", "--num-frames 1 -num-sensors 8 --lidar-type Rotary")
+        python_sample_test("tests-standalone_benchmarks-benchmark_rtx_lidar_solid_state", "standalone_examples/benchmarks/benchmark_rtx_lidar.py", "--num-frames 1 -num-sensors 8 --lidar-type Solid_State")
+        python_sample_test("tests-standalone_benchmarks-benchmark_sdg_simple", "standalone_examples/benchmarks/benchmark_sdg.py", "--num-frames 1 --num-cameras 2 --resolution 1280 720 --asset-count 10 --annotators rgb distance_to_camera --disable-viewport-rendering --delete-data-when-done --headless --print-results")
+        python_sample_test("tests-standalone_benchmarks-benchmark_sdg_advanced", "standalone_examples/benchmarks/benchmark_sdg.py", "--num-frames 1 --num-cameras 2 --resolution 1280 720 --asset-count 10 --annotators all --disable-viewport-rendering --delete-data-when-done --headless --print-results")
 
     if os.target() == "linux" then
     group "docker_tests"