+++ source/extensions/omni.isaac.benchmarks/premake5.lua
@@ -1,7 +1,5 @@
 local ext = get_current_extension_info()
-project_ext (ext, { 
-    define_test = false
-})
+project_ext (ext)
 
 repo_build.prebuild_link {
     { "data", ext.target_dir.."/data" },