+++ source/standalone_examples/benchmarks/benchmark_robots_nova_carter_ros2.py
@@ -10,7 +10,7 @@
 import argparse
 
 parser = argparse.ArgumentParser()
-parser.add_argument("-n", "--n-robot", type=int, default=1, help="Number of robots")
+parser.add_argument("--num-robots", type=int, default=1, help="Number of robots")
 parser.add_argument(
     "--enable-3d-lidar", type=int, default=0, choices=range(0, 1 + 1), help="Number of 3D lidars to enable, per robot."
 )
@@ -25,7 +25,7 @@ parser.add_argument(
     help="Number of Hawk camera stereo pairs to enable, per robot.",
 )
 parser.add_argument("--num-gpus", type=int, default=1, help="Number of GPUs on machine.")
-parser.add_argument("--test", default=False, action="store_true", help="Run in test mode")
+parser.add_argument("--num-frames", type=int, default=600, help="Number of frames to run benchmark for")
 parser.add_argument(
     "--backend-type",
     default="OsmoKPIFile",
@@ -35,21 +35,18 @@ parser.add_argument(
 
 args, unknown = parser.parse_known_args()
 
-n_robot = args.n_robot
+n_robot = args.num_robots
 enable_3d_lidar = args.enable_3d_lidar
 enable_2d_lidar = args.enable_2d_lidar
 enable_hawks = args.enable_hawks
 n_gpu = args.num_gpus
+n_frames = args.num_frames
 
 import numpy as np
 from isaacsim import SimulationApp
 
 simulation_app = SimulationApp({"headless": True, "max_gpu_count": n_gpu})
 
-TEST_NUM_APP_UPDATES = 60 * 10
-
-if args.test:
-    TEST_NUM_APP_UPDATES = 1
 
 import carb
 import omni
@@ -78,7 +75,7 @@ benchmark = BaseIsaacBenchmark(
     },
     backend_type=args.backend_type,
 )
-benchmark.set_phase("loading")
+benchmark.set_phase("loading", start_recording_frametime=False, start_recording_runtime=True)
 
 enable_extension("omni.isaac.ros2_bridge")
 omni.kit.app.get_app().update()
@@ -164,7 +161,7 @@ benchmark.store_measurements()
 benchmark.set_phase("benchmark")
 
 
-for _ in range(1 if benchmark.test_mode else TEST_NUM_APP_UPDATES):
+for _ in range(1, n_frames):
     omni.kit.app.get_app().update()
 
 benchmark.store_measurements()